requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyyaml==6.0.1
rich==13.7.1
prompt_toolkit==3.0.47
rapidfuzz==3.5.2
playwright==1.45.0
ollama==0.1.8
pytest==8.2.2
